requirements.txt
numpy==1.21.4
opencv-python-headless==4.5.4.60
ffmpegcv==0.2.3
gpxpy==1.5.0
tqdm==4.62.3
piexif==1.1.3
matplotlib==3.5.1
pandas==1.3.5
geopandas==0.10.2
Pillow==8.4.0
git+https://github.com/localdevices/py360convert.git
